// src/types/word-adapter.ts
export class WordAdapter {
    id?: number;
    audioURL?: string;
    wordId?: number;
    wordHeadId?: number;
    wordRootId?: number;
    wordRootDefinitionId?: number;
    wordInstanceId?: number;
    wordUsageCount?: number;
    headUsageCount?: number;
    rootUsageCount?: number;
    instanceUsageCount?: number;
    headPick?: boolean;
    rootPick?: boolean;
    sharedMeaningId?: number;
}